+++ src/Platform/Fields/Builder.php
@@ -3,7 +3,6 @@
 namespace Orchid\Platform\Fields;
 
 use Orchid\Platform\Screen\Repository;
-use Orchid\Platform\Exceptions\TypeException;
 
 class Builder
 {
@@ -35,14 +34,26 @@ class Builder
     /**
      * Builder constructor.
      *
-     * @param array  $fields
-     * @param        $data
-     * @param string $language
-     * @param string $prefix
+     * @param array       $fields
+     * @param             $data
+     * @param string|null $language
+     * @param string|null $prefix
+     *
+     * @throws \Orchid\Platform\Exceptions\TypeException
      */
     public function __construct(array $fields, $data, string $language = null, string $prefix = null)
     {
-        $this->fields = Parser::parseFields($fields);
+        //deprecated
+        foreach ($fields as $key => $item){
+            if(!is_object($item)){
+
+                $RawParse = Parser::parseFields([$item]);
+                $RawField = array_shift($RawParse)->toArray();
+                $fields[$key] =  Field::make($RawField);
+            }
+        }
+
+        $this->fields = $fields;
         $this->data = $data ?? $data = new Repository([]);
 
         $this->language = $language;
@@ -77,32 +88,21 @@ class Builder
      * Generate a ready-made html form for display to the user.
      *
      * @return string
-     * @throws TypeException
      */
     public function generateForm() : string
     {
         $fields = $this->fields;
-        $availableFormFields = [];
 
         $this->form = '';
-        foreach ($fields as $field => $config) {
-            $fieldClass = config('platform.fields.'.$config['tag']);
+        foreach ($fields as $field) {
 
-            if (is_null($fieldClass)) {
-                throw new TypeException('Field '.$config['tag'].' does not exist');
+            $field->set('lang',$this->language);
+            $field->set('prefix',$this->buildPrefix($field));
+
+            foreach ($this->fill($field->getAttributes()) as $key => $value){
+                $field->set($key,$value);
             }
 
-            $config['lang'] = $this->language;
-            $config['prefix'] = $this->buildPrefix($config);
-            $config = $this->fill($config);
-
-            $firstTimeRender = false;
-            if (! in_array($fieldClass, $availableFormFields)) {
-                array_push($availableFormFields, $fieldClass);
-                $firstTimeRender = true;
-            }
-
-            $field = (new $fieldClass())->create($config, $firstTimeRender);
             $this->form .= $field->render();
         }
 
@@ -110,58 +110,58 @@ class Builder
     }
 
     /**
-     * @param $config
+     * @param $field
      *
      * @return string
      */
-    private function buildPrefix($config)
+    private function buildPrefix($field)
     {
-        if (isset($config['prefix'])) {
-            $prefixArray = array_filter(explode(' ', $config['prefix']));
+        $prefix = $field->get('prefix',null);
 
-            foreach ($prefixArray as $prefix) {
-                $config['prefix'] .= '['.$prefix.']';
+        if (!is_null($prefix)) {
+
+            foreach (array_filter(explode(' ', $prefix)) as $name) {
+                $prefix .= '['.$name.']';
             }
-
-            return $config['prefix'];
+            return $prefix;
         }
 
         return $this->prefix;
     }
 
     /**
-     * @param $config
+     * @param $attributes
      *
      * @return mixed
      */
-    private function fill($config)
+    private function fill($attributes)
     {
-        $name = array_filter(explode(' ', $config['name']));
+        $name = array_filter(explode(' ', $attributes['name']));
         $name = array_shift($name);
 
-        $config['value'] = $this->getValue($name, $config['value'] ?? null);
+        $attributes['value'] = $this->getValue($name, $attributes['value'] ?? null);
 
         $binding = explode('.', $name);
         if (! is_array($binding)) {
-            return $config;
+            return $attributes;
         }
 
-        $config['name'] = '';
+        $attributes['name'] = '';
         foreach ($binding as $key => $bind) {
-            if (! is_null($config['prefix'])) {
-                $config['name'] .= '['.$bind.']';
+            if (! is_null($attributes['prefix'])) {
+                $attributes['name'] .= '['.$bind.']';
                 continue;
             }
 
             if ($key === 0) {
-                $config['name'] .= $bind;
+                $attributes['name'] .= $bind;
                 continue;
             }
 
-            $config['name'] .= '['.$bind.']';
+            $attributes['name'] .= '['.$bind.']';
         }
 
-        return $config;
+        return $attributes;
     }
 
     /**